+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v0.13.0 - Unreleased
 
+### Added
+
+- Added dotnet-isolated support for Embeddings
+
 ### Breaking changes
 
 - Changed default embeddings model from `text-embedding-ada-002` to `text-embedding-3-small`